requirements.txt
Flask
flask-cors
nibabel
numpy
Pillow
scikit-image
trimesh
matplotlib
torch
monai
